move hc_* file functions from shared.c/.h to filehandling.c/.h
