+++ include/shared.h
@@ -14,7 +14,9 @@
 #include <fcntl.h>
 #include <ctype.h>
 #include <math.h>
+
 #include "zlib.h"
+#include "filehandling.h"
 
 #if defined (_WIN)
 #include <winsock2.h> // needed for select()
@@ -61,27 +63,6 @@ bool hc_string_is_digit (const char *s);
 
 void hc_string_trim_trailing (char *s);
 void hc_string_trim_leading (char *s);
-
-#if defined (__CYGWIN__)
-int    _wopen(const char *path, int oflag, ...);
-#endif
-
-bool   hc_fopen    (HCFILE *fp, const char *path, char *mode);
-int    hc_fscanf   (HCFILE *fp, const char *format, void *ptr);
-int    hc_fprintf  (HCFILE *fp, const char *format, ...);
-int    hc_vfprintf (HCFILE *fp, const char *format, va_list ap);
-int    hc_fseek    (HCFILE *fp, off_t offset, int whence);
-void   hc_rewind   (HCFILE *fp);
-off_t  hc_ftell    (HCFILE *fp);
-int    hc_fgetc    (HCFILE *fp);
-int    hc_fileno   (HCFILE *fp);
-int    hc_feof     (HCFILE *fp);
-void   hc_fflush   (HCFILE *fp);
-void   hc_fclose   (HCFILE *fp);
-int    hc_fputc    (int c, HCFILE *fp);
-char  *hc_fgets    (char *buf, int len, HCFILE *fp);
-size_t hc_fwrite   (void *ptr, size_t size, size_t nmemb, HCFILE *fp);
-size_t hc_fread    (void *ptr, size_t size, size_t nmemb, HCFILE *fp);
 
 bool hc_same_files (char *file1, char *file2);
